+++ sacred/observers/mongo.py
@@ -234,14 +234,14 @@ class MongoObserver(RunObserver):
         self.run_entry['resources'].append((filename, md5hash))
         self.save()
 
-    def artifact_event(self, filename):
+    def artifact_event(self, name, filename):
         with open(filename, 'rb') as f:
-            _, tail = os.path.split(filename)
             run_id = self.run_entry['_id']
-            db_filename = 'artifact://{}/{}/{}'.format(
-                self.run_entry['experiment']['name'], run_id, tail)
+            db_filename = 'artifact://{}/{}/{}'.format(self.runs.name, run_id,
+                                                       name)
             file_id = self.fs.put(f, filename=db_filename)
-        self.run_entry['artifacts'].append(file_id)
+        self.run_entry['artifacts'].append({'name': name,
+                                            'file_id': file_id})
         self.save()
 
     def __eq__(self, other):